% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_model.R
\name{generate_model}
\alias{generate_model}
\title{This function generates a model specified by the user}
\usage{
generate_model(x_names, c_names, model, N, rt_limits)
}
\description{
This function generates a model specified by the user
}
