% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regddmfit_class.R
\name{get_stan_fit}
\alias{get_stan_fit}
\title{Get the rstan fit of regddmfit objects}
\usage{
get_stan_fit(fit)
}
\arguments{
\item{fit}{A \code{regddmfit} object to summary}
}
\value{
A \code{stanfit} object.
}
\description{
Get the \code{\link[rstan]{stanfit}} object of the \code{\link{regddmfit}}
object to perform further analysis and diagnosis.
}
