% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regddmfit_class.R
\name{print.regddmfit}
\alias{print.regddmfit}
\title{Printed summary of \pkg{RegDDM} fit object}
\usage{
\method{print}{regddmfit}(x, digits = 3, ...)
}
\arguments{
\item{x}{An \code{\link{regddmfit}} object to print}

\item{digits}{digits of the output results. Default value is 3.}

\item{...}{Unused...}
}
\value{
No values are returned.
}
\description{
Summarize the posterior distributions of estimated regression coefficients
and print necessary information.
}
\seealso{
\code{\link{summary.regddmfit}} Table summaries of the \code{regddmfit} object.
}
