% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regddm_data}
\alias{regddm_data}
\title{Example of LS task dataset from our previous Cognitive Reserve study}
\format{
A list with two items
\describe{
\item{data1}{Subject-level dataframe with 49 rows and 4 columns: \code{id}, \code{y}, \code{age} and \code{education}}
\item{data2}{Trial-level dataframe with 6032 rows and 4 columns: \code{id}, \code{memload}, \code{response}, \code{rt}}
}
}
\usage{
regddm_data
}
\description{
This is a dataset with 49 subjects and 6032 trials. Each trial, the subject
is presented with different number of letters (\code{memload} variable). After 7
seconds, the subect is presented with another letter, and asked to decide
whether the letter is in the letter set at the beginning. The response (correct = 1
, incorrect = 0) and response time (\code{rt}, in seconds) is recorded. The age
and education of the subject (both in years) are also recorded. \code{y} is the
IQ of the subject estimated using the American NART.
}
\examples{
data("regddm_data")
head(regddm_data$data1)
head(regddm_data$data2)
}
\keyword{datasets}
