% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regddm_tutorial}
\alias{regddm_tutorial}
\title{Simulated example dataset}
\format{
An object of class \code{list} of length 2.
}
\usage{
regddm_tutorial
}
\description{
This is the same simulated dataset used in the tutorial part of the RegDDM paper.

\eqn{v = x_1\times v_{x_1}}

\eqn{y \sim N(1 + 1\times v_{x_1} + 1\times c_1, 1)}
}
\keyword{datasets}
