% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regddmfit_class.R
\name{summary.regddmfit}
\alias{summary.regddmfit}
\title{Summary of \pkg{RegDDM} fit object}
\usage{
\method{summary}{regddmfit}(object, ...)
}
\arguments{
\item{object}{A \code{\link{regddmfit}} object to summary}

\item{...}{parameters passed to \code{\link[rstan]{summary,stanfit-method}}}
}
\value{
The summary method returns a named list of \code{glm_coefficiets},
\code{subject_ddm_param}, \code{group_param}, and \code{missing_value}.
Each element is a \code{\link[tibble]{tibble}} data frame of posterior
summary statistics of the regression coefficient, DDM parameter of each
subject, group mean and standard deviation of DDM parameters and covariates,
plus the estimated missing values.
}
\description{
Summarize the posterior distributions of estimated parameters and group them into
four categories.
}
\seealso{
\code{\link{print.regddmfit}} a printed summary of the \code{regddmfit} object.
}
