% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_fake_data.R
\name{generate_fake_data}
\alias{generate_fake_data}
\title{This function generates a fake dataset using easyRT}
\usage{
generate_fake_data(
  N = 10,
  beta_0 = 0.2,
  beta_c1 = -1,
  beta_c2 = 0,
  beta_v_x1 = 1,
  beta_v_x2 = 0,
  sigma_y = 0.1,
  rep_trial = 4
)
}
\description{
This function generates a fake dataset using easyRT
}
