% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegDDM.R
\name{regddm}
\alias{regddm}
\title{Main function of RegDDM}
\usage{
regddm(
  data1,
  data2,
  model = default_model,
  init = "default",
  scale = FALSE,
  chains = 4,
  cores = 4,
  thin = 1,
  ...
)
}
\description{
Main function of RegDDM
}
