% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regddm_data}
\alias{regddm_data}
\title{Example dataset used from Cognitive reserve study.}
\format{
An object of class \code{list} of length 2.
}
\usage{
regddm_data
}
\description{
Tbd
}
\keyword{datasets}
