% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regddm_tutorial}
\alias{regddm_tutorial}
\title{Simulated example dataset}
\format{
An object of class \code{list} of length 2.
}
\usage{
regddm_tutorial
}
\description{
This is the same simulated dataset used in the tutorial part of the RegDDM paper.
v = x1\*v_x1
y ~ N(1 + 1\*v_x1 + 1\*c1, 1)
}
\keyword{datasets}
