% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regddmfit_class.R
\name{regddmfit}
\alias{regddmfit}
\title{Class of models fitted by \pkg{RegDDM}}
\usage{
regddmfit(data1, data2, model, family, stan_fit)
}
\description{
\code{regddmfit} is an S3 object sotring the fitted models of \pkg{RegDDM}.
It contains information used to fit the model and the resulting
\link[rstan]{\code{stanfit}} and can be summarized and printed using
\code{\link{summary.regddmfit}} and \code{\link{print.redggmfit}}.
}
\details{
Use \code{methods(class = "regddmfit")} for a list of available methods.
}
\section{Slots}{

\describe{
\item{\code{data1}}{Subject-level data frame.}

\item{\code{data2}}{Trial-level data frame.}

\item{\code{model}}{A list containing 0-5 formulas, specifying the dependence
structure between variables.}

\item{\code{family}}{Family of distribution of the outcome.}

\item{\code{stan_fit}}{Fitted \code{stan} model.}
}}

